project-{topic}-pro/backend/requirements.txt
```
fastapi>=0.104.1
uvicorn[standard]>=0.23.2
sqlalchemy>=2.0.22
asyncpg>=0.28.0
psycopg2-binary>=2.9.9 # For Alembic's sync operations if needed, or local development
pydantic>=2.4.2
pydantic-settings>=2.0.3
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
alembic>=1.12.0
python-dotenv>=1.0.0
redis>=5.0.1 # For redis.asyncio
fastapi-cache2>=0.2.1
fastapi-limiter>=0.1.5
pytest>=7.4.3
pytest-asyncio>=0.21.1
httpx>=0.25.1
locust>=2.19.0
```